requirements.txt
django>=1.5.1
django-environ>=0.3.0
wheel==0.24.0
python-kong>=0.1.11
pytz>=2015.4
